+++ include/entities/nyannyan.hpp
@@ -4,7 +4,7 @@
 class NyanNyan : public Enemy {
 public:
 
-  NyanNyan(float x, float y, Renderer& r, Player* p) : Enemy(x,y,r,p) { 
+  NyanNyan(float x, float y, Renderer& r, Player* p) : Enemy(x,y,r,p, EntityType::NyanNyan) { 
     state = State::Idle;
 
     facing = Facing::Left;
@@ -19,5 +19,8 @@ public:
   //virtual void update(float dt, const std::vector<Platform>& platforms) override;
 
   void ai(float dt, const std::vector<Platform>& platforms) override;
+
+  EntityType getEntityTypeEnum() const override { return EntityType::NyanNyan; }
+  const char* getEntityType() const override { return "NyanNyan"; }
 
 };